nvme-print-stdout: Refactor stdout print code to use print_ops
Use the newly introduced print_ops infrastructure to for the stdout
printing code.
