+++ meson.build
@@ -245,6 +245,7 @@ sources = [
   'nvme.c',
   'nvme-models.c',
   'nvme-print.c',
+  'nvme-print-stdout.c',
   'nvme-rpmb.c',
   'nvme-wrap.c',
   'plugin.c',
