+++ nvme-print.h
@@ -106,6 +106,8 @@ static inline struct print_ops *nvme_get_json_print_ops(enum nvme_print_flags fl
 
 #endif /* !CONFIG_JSONC */
 
+struct print_ops *nvme_get_stdout_print_ops(enum nvme_print_flags flags);
+
 void nvme_show_status(int status);
 void nvme_show_lba_status_info(__u32 result);
 void nvme_show_relatives(const char *name);
